Move remaining .d.ts files to .ts
