+++ packages/desktop-client/globals.ts
